Issue 32 renamed canisctl to sirius (#34)
