+++ cmd/sirius/main.go
@@ -7,7 +7,7 @@ SPDX-License-Identifier: Apache-2.0
 package main
 
 import (
-	"github.com/scoir/canis/pkg/canisctl/cmd"
+	"github.com/scoir/canis/pkg/sirius/cmd"
 )
 
 func main() {
