+++ docker/canis/Dockerfile
@@ -6,7 +6,7 @@ RUN DEBIAN_FRONTEND=noninteractive apt-get install -y tzdata
 
 COPY ./bin/steward /usr/local/bin/steward
 COPY ./bin/agent /usr/local/bin/agent
-COPY ./bin/canisctl /usr/local/bin/canisctl
+COPY ./bin/sirius /usr/local/bin/sirius
 
 WORKDIR /usr/local/canis
 COPY ./static ./static
