+++ pkg/sirius/cmd/start.go
