+++ pkg/sirius/cmd/status.go
