+++ pkg/sirius/cmd/stop.go
